--- src/component/BurgerMenu.css
@@ -1,66 +1,0 @@
-.bm-burger-button {
-    position: fixed;
-    width: 36px;
-    height: 20px;
-    left: 25px;
-    top: 25px;
-}
-
-/* Color/shape of burger icon bars */
-.bm-burger-bars {
-    background: #2c3e50;
-}
-
-/* Color/shape of burger icon bars on hover*/
-.bm-burger-bars-hover {
-    background: #667db6;
-}
-
-/* Position and sizing of clickable cross button */
-.bm-cross-button {
-    height: 24px;
-    width: 24px;
-    background-color: black;
-}
-
-/* Color/shape of close button cross */
-.bm-cross {
-    background: #bdc3c7;
-}
-
-/*
-Sidebar wrapper styles
-Note: Beware of modifying this element as it can break the animations - you should not need to touch it in most cases
-*/
-.bm-menu-wrap {
-    position: fixed;
-    height: 100%;
-}
-
-/* General sidebar styles */
-.bm-menu {
-    background: #2c3e50;
-    padding: 2.5em 1.5em 0;
-    font-size: 1.15em;
-}
-
-/* Morph shape necessary with bubble or elastic */
-.bm-morph-shape {
-    fill: #373a47;
-}
-
-/* Wrapper for item list */
-.bm-item-list a {
-    color: #b8b7ad;
-    padding: 0.8em;
-}
-
-/* Individual item */
-.bm-item {
-    display: inline-block;
-}
-
-/* Styling of overlay */
-.bm-overlay {
-    background: rgba(0, 0, 0, 0.3);
-}
