--- src/component/Main.js
@@ -1,21 +1,0 @@
-import React from 'react';
-import { Switch, Route } from 'react-router-dom';
-
-import LandingPage from './LandingPage';
-import AboutMe from './Aboutme';
-import Contact from './Contact';
-import Projects from './Project';
-import Resume from './Resume';
-
-
-const Main = () => (
-    <Switch>
-        <Route exact path="/" component={LandingPage} />
-        <Route path="/aboutme" component={AboutMe} />
-        <Route path="/contact" component={Contact} />
-        <Route path="/projects" component={Projects} />
-        <Route path="/resume" component={Resume} />
-    </Switch>
-);
-
-export default Main;
